Add Update, Delete method to WorkoutController. Change id from int to long.

+++ src/main/java/be/thomasmol/workout_tracker/workout/Workout.java
@@ -13,7 +13,7 @@ public class Workout {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
 
     @Column(name = "name")
     private String name;
@@ -35,7 +35,7 @@ public class Workout {
     public Workout() {
     }
 
-    public Workout(int id, String name, LocalDateTime dateRecorded, int durationInMinutes, Set<Exercise> exercises) {
+    public Workout(long id, String name, LocalDateTime dateRecorded, int durationInMinutes, Set<Exercise> exercises) {
         this.id = id;
         this.name = name;
         this.dateRecorded = dateRecorded;
@@ -43,11 +43,11 @@ public class Workout {
         this.exercises = exercises;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
